+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore(identity-center): update sdk v3
 - chore(route53resolver): update sdk v3
 - chore(s3): update sdk v3
+- chore(organizations): update sdk v3
 
 ### Configuration Changes
 - chore(config): add s3 object level and additional network controls